requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
httpx==0.27.0
anthropic==0.40.0
gspread==6.1.2
google-auth==2.36.0
python-dotenv==1.0.0
pydantic-settings==2.6.0
